+++ timeline/templatetags/timeline_tags.py
@@ -4,15 +4,18 @@ from django import template
 from django.urls import reverse
 from django.utils.timezone import localtime
 
-# Projectroles dependency
-from projectroles.plugins import get_app_plugin
+from djangoplugins.models import Plugin
 
 from timeline.api import TimelineAPI
 from timeline.models import ProjectEvent
 
+
 timeline = TimelineAPI()
 register = template.Library()
 
+
+ICON_PROJECTROLES = 'mdi:cube'
+ICON_UNKNOWN_APP = 'mdi:help-circle'
 STATUS_STYLES = {
     'OK': 'bg-success',
     'INIT': 'bg-secondary',
@@ -39,16 +42,51 @@ def get_event_description(event, request=None):
 
 
 @register.simple_tag
-def get_details_events(project, view_classified):
+def get_details_events(project, view_classified=False):
     """Return recent events for card on project details page"""
     events = ProjectEvent.objects.filter(project=project)
-
     if not view_classified:
         events = events.exclude(classified=True)
-
     events = events.order_by('-pk')
-
     return [x for x in events if x.get_current_status().status_type == 'OK'][:5]
+
+
+@register.simple_tag
+def get_plugin_lookup():
+    """Return lookup dict of app plugins with app name as key"""
+    return {p.name: p.get_plugin() for p in Plugin.objects.all()}
+
+
+@register.simple_tag
+def get_app_icon_html(plugin_lookup, event):
+    """Return icon link HTML for app by plugin lookup"""
+    url = None
+    url_kwargs = {}
+    if event.project:
+        url_kwargs['project'] = event.project.sodar_uuid
+    title = event.app
+    icon = ICON_UNKNOWN_APP  # Default in case the plugin is not found
+
+    if event.app == 'projectroles':
+        if event.project:
+            url = reverse('projectroles:detail', kwargs=url_kwargs)
+        title = 'Projectroles'
+        icon = ICON_PROJECTROLES
+    elif event.app in plugin_lookup.keys():
+        plugin = plugin_lookup[event.app]
+        entry_point = getattr(plugin, 'entry_point_url_id', None)
+        if entry_point:
+            url = reverse(entry_point, kwargs=url_kwargs)
+        title = plugin.title
+        if getattr(plugin, 'icon', None):
+            icon = plugin.icon
+
+    return (
+        '<a {} title="{}" data-toggle="tooltip" data-placement="top">'
+        '<i class="iconify" data-icon="{}"></i></a>'.format(
+            'href="{}"'.format(url) if url else '', title, icon
+        )
+    )
 
 
 # Template rendering -----------------------------------------------------------
@@ -62,30 +100,6 @@ def get_status_style(status):
         if status.status_type in STATUS_STYLES
         else 'bg-light'
     )
-
-
-@register.simple_tag
-def get_app_url(event):
-    """Return URL for event application"""
-    url_kwargs = {}
-    if event.project:
-        url_kwargs['project'] = event.project.sodar_uuid
-
-    # Projectroles is a special case
-    if event.app == 'projectroles' and event.project:
-        return reverse('projectroles:detail', kwargs=url_kwargs)
-    elif event.app != 'projectroles':
-        app_plugin = get_app_plugin(event.app, plugin_type='project_app')
-        if app_plugin:
-            return reverse(
-                app_plugin.entry_point_url_id,
-                kwargs=url_kwargs,
-            )
-        # Try site apps
-        app_plugin = get_app_plugin(event.app, plugin_type='site_app')
-        if app_plugin:
-            return reverse(app_plugin.entry_point_url_id)
-    return '#'
 
 
 @register.simple_tag
@@ -203,28 +217,23 @@ def html_print_array(array, str_list, indent):
 
 
 @register.filter
-def collect_extra_data(event: ProjectEvent):
+def collect_extra_data(event):
     ls = []
-
     if event.extra_data is not None and len(event.extra_data) > 0:
         ls.append(('extra-data', 'Extra Data', event))
-
     for status in event.get_status_changes():
         if status.extra_data is not None and len(status.extra_data) > 0:
             ls.append(
                 ('status-extra-data', 'Status: ' + status.status_type, status)
             )
-
     return ls
 
 
 @register.filter
-def has_extra_data(event: ProjectEvent):
+def has_extra_data(event):
     if event.extra_data is not None and len(event.extra_data) > 0:
         return True
-
     for status in event.get_status_changes():
         if status.extra_data is not None and len(status.extra_data) > 0:
             return True
-
     return False